# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 回测平台 API
  version: 1.0.0
  description: |
    回测平台提供组合回测、资产分析、蒙特卡洛模拟、组合优化、有效前沿计算、
    战术分配、信号分析、PCA 主成分分析、LETF 滑点分析、目标优化等量化投资工具。

    ## 认证
    - 管理端点（admin、data/manage）需要 `x-api-key` 请求头认证
    - 计算密集型端点（backtest、backtest-optimizer）可选 `x-api-key`，有 Key 时验证身份，无 Key 时放行但记录匿名访问
    - 健康检查和 metrics 端点不需要认证

    ## 速率限制
    - 普通 API：100 次/15 分钟/IP
    - 计算密集型 API（backtest、backtest-optimizer）：10 次/分钟/IP
servers:
  - url: http://localhost:5001/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌，通过 /auth/login 获取
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          oneOf:
            - type: string
            - type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: 机器可读错误码
                message:
                  type: string
                  description: 人类可读错误描述

    # ===== 认证鉴权 =====
    # 企业理由：API 契约文档是前后端协作的唯一可靠依据。
    # 认证流程涉及令牌签发、刷新、撤销，任何字段歧义都会导致集成故障，
    # 因此必须将 LoginRequest / TokenResponse / UserInfo 等结构固化到 OpenAPI 规范中，
    # 确保客户端与服务端对令牌生命周期管理达成一致理解。
...
    Asset:
      type: object
      required:
        - ticker
        - weight
      properties:
        id:
          type: string
        ticker:
          type: string
          description: 资产代码
        weight:
          type: number
          description: 权重百分比
    RebalanceBands:
      type: object
      properties:
        enabled:
          type: boolean
        absoluteBand:
          type: number
          description: 绝对偏离百分比
        relativeBand:
          type: number
          description: 相对偏离百分比
        upperBand:
          type: number
          description: 上限偏离百分比
        lowerBand:
          type: number
          description: 下限偏离百分比
    Portfolio:
      type: object
      required:
        - id
        - name
        - assets
        - rebalanceFrequency
      properties:
        id:
          type: string
        name:
          type: string
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        rebalanceFrequency:
          type: string
          enum: [daily, weekly, monthly, quarterly, annual, none, threshold]
        rebalanceThreshold:
          type: number
          description: 偏离调仓阈值（百分比），仅 threshold 模式有效
        rebalanceOffset:
          type: number
          description: 调仓偏移（交易日）
        rebalanceBands:
          $ref: '#/components/schemas/RebalanceBands'
        drag:
          type: number
          description: 年化拖累百分比，如 0.5 表示每年额外扣除 0.5%
        totalReturn:
          type: boolean
          description: 是否将分红再投资
        isGlidepath:
          type: boolean
        glidepathFrom:
          type: string
        glidepathTo:
          type: string
        glidepathYears:
          type: number
        glidepathToWeights:
          type: array
          items:
            type: number
    CashflowLeg:
      type: object
      required:
        - id
        - amount
        - type
        - frequency
        - offset
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [contribution, withdrawal]
        frequency:
          type: string
          enum: [yearly, monthly, quarterly, weekly]
        offset:
          type: number
        until:
          type: string
          description: 结束日期
...
    BacktestParameters:
      type: object
      required:
        - startDate
        - endDate
        - startingValue
        - adjustForInflation
        - rollingWindowMonths
        - benchmarkTicker
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        startingValue:
          type: number
        baseCurrency:
          type: string
          enum: [usd, cny]
          default: usd
        adjustForInflation:
          type: boolean
        rollingWindowMonths:
          type: number
          default: 12
        benchmarkTicker:
          type: string
        extendedWithdrawalStats:
          type: boolean
        cashflowLegs:
          type: array
          items:
            $ref: '#/components/schemas/CashflowLeg'
        oneTimeCashflows:
          type: array
          items:
            $ref: '#/components/schemas/OneTimeCashflow'

    # ===== 统计 =====
...
    BacktestResult:
      type: object
      properties:
        portfolios:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioResult'
        correlations:
          type: array
          items:
            type: array
            items:
              type: number
        benchmarkGrowth:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              value:
                type: number

    # ===== 战术分配 =====
...
    OptimizeRequest:
      type: object
      required:
        - tickers
        - parameters
      properties:
        tickers:
          type: array
          items:
            type: string
          maxItems: 50
        objective:
          type: string
          enum: [maxSharpe, minVolatility, maxReturn]
          description: 优化目标
        constraints:
          type: object
          properties:
            minWeight:
              type: number
            maxWeight:
              type: number
        parameters:
          $ref: '#/components/schemas/BacktestParameters'
        riskFreeRate:
          type: number
        numIterations:
          type: number
          description: 迭代次数，上限 100000，默认 10000
...
    MonteCarloParams:
      type: object
      properties:
        numSimulations:
          type: number
          description: 模拟次数
        blockSize:
          type: number
          description: 区块大小
        withReplacement:
          type: boolean
        confidenceLevel:
          type: number
        distribution:
          type: string
        seed:
          type: number
paths:
...
  /backtest/search:
    get:
      tags: [backtest]
      summary: 搜索资产代码
      description: 根据关键词搜索匹配的资产代码
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: 返回结果数量上限
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '400':
          description: 缺少 query 参数
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 搜索失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /backtest/portfolio:
    post:
      tags: [backtest]
      summary: 运行组合回测
      description: |
        接收组合配置和回测参数，运行组合回测。
        优先使用 Rust 引擎计算，失败时降级到 Node.js 引擎。
        降级模式下响应会包含 `degraded: true` 和 `degradedWarning` 字段。
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - portfolios
                - parameters
              properties:
                portfolios:
                  type: array
                  items:
                    $ref: '#/components/schemas/Portfolio'
                  maxItems: 50
                parameters:
                  $ref: '#/components/schemas/BacktestParameters'
      responses:
        '200':
          description: 回测结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/BacktestResult'
                  degraded:
                    type: boolean
                    description: 是否降级到 Node.js 引擎
                  degradedWarning:
                    type: string
                  warnings:
                    type: array
                    items:
                      type: string
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 回测运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /backtest/analysis:
    post:
      tags: [backtest]
      summary: 运行资产分析
      description: 对多个资产行情数据进行分析，计算收益率、相关性等统计指标
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tickers
                - parameters
              properties:
                tickers:
                  oneOf:
                    - type: array
                      items:
                        type: string
                      maxItems: 50
                    - type: string
                      description: 空格或逗号分隔的 ticker 字符串
                parameters:
                  $ref: '#/components/schemas/BacktestParameters'
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tickers:
                        type: array
                        items:
                          type: object
                      correlations:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 分析运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /backtest/monte-carlo:
    post:
      tags: [backtest]
      summary: 运行蒙特卡洛模拟
      description: 对投资组合运行蒙特卡洛模拟，支持单个或多个组合
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                portfolio:
                  $ref: '#/components/schemas/Portfolio'
                  description: 单个组合（与 portfolios 二选一）
                portfolios:
                  type: array
                  items:
                    $ref: '#/components/schemas/Portfolio'
                  maxItems: 50
                  description: 多个组合（与 portfolio 二选一）
                parameters:
                  $ref: '#/components/schemas/BacktestParameters'
                mcParams:
                  $ref: '#/components/schemas/MonteCarloParams'
      responses:
        '200':
          description: 蒙特卡洛模拟结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 模拟运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /backtest/optimize:
    post:
      tags: [backtest]
      summary: 运行组合优化
      description: 根据目标函数和约束条件优化资产权重
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeRequest'
      responses:
        '200':
          description: 优化结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 优化运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /backtest/efficient-frontier:
    post:
      tags: [backtest]
      summary: 计算有效前沿
      description: 计算给定资产集合的有效前沿曲线
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tickers
                - parameters
              properties:
                tickers:
                  type: array
                  items:
                    type: string
                  maxItems: 50
                numPoints:
                  type: number
                  description: 有效前沿上的点数，默认 20
                  default: 20
                parameters:
                  $ref: '#/components/schemas/BacktestParameters'
                riskFreeRate:
                  type: number
                  default: 0.02
                numIterations:
                  type: number
      responses:
        '200':
          description: 有效前沿结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 计算运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== 数据查询 =====
...